allow multiple verification types
